requirements.txt
beautifulsoup4==4.12.3
matplotlib==3.7.2
networkx==3.1
openpyxl==3.0.10
pandas==2.2.1
python_dateutil==2.8.2
Requests==2.31.0
tqdm==4.65.0
